requirements.txt
python==3.10.9
tensorflow==2.12.0
keras==2.12.0
numpy>=1.21,<1.24
protobuf==3.20.3

opencv-python-headless>=4.6
pillow>=9.0
matplotlib>=3.5

scikit-learn>=1.0

streamlit==1.25.0
streamlit-webrtc>=0.45.0
av

pyyaml